requirements.txt
fastapi
uvicorn
jinja2
python-multipart
python-dotenv
sqlmodel
requests
pandas
passlib[bcrypt]
python-jose[cryptography]
huggingface_hub
pypdf2
